[NEW] #62 - finished majority of the enviromental variables, still a some issues that need to be solved
