+++ sources/graph/interface/add_argv.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   add_argv.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: schiper <schiper@student.42.fr>            +#+  +:+       +#+        */
+/*   By: iatilla- <iatilla-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/25 18:09:52 by schiper           #+#    #+#             */
-/*   Updated: 2025/04/29 14:15:28 by schiper          ###   ########.fr       */
+/*   Updated: 2025/05/01 02:31:12 by iatilla-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,6 +25,7 @@ static int	copy_data(char ***new_argv, char **old_argv, unsigned int len)
 	}
 	return (0);
 }
+
 // THIS FUNCTION MIGHT BE FULL COOKED -> GIGA LEAKS IN FUTURE INCOMING
 void	add_argv(t_cmd **cmd, t_parsed_data **data)
 {
